(chore) move more SaturationFilter subcomponent buttons to FilterButtons file; put FilterStep enum in types file
