+++ frontend/src/components/SaturationFilter/FilterButtons.tsx
@@ -1,3 +1,4 @@
+import { BiArrowToLeft, BiArrowToRight } from "react-icons/bi";
 import { operators } from "./helpers/operators";
 import * as S from "./SaturationFilter.style";
 
@@ -28,5 +29,40 @@ export default function OperatorButton({ operator, onClick }: OperatorButtonProp
 		<S.OperatorButton key={operator} value={operator} onClick={() => onClick?.()}>
 			<span>{operator}</span>
 		</S.OperatorButton>
+	);
+}
+
+type ThresholdButtonProps = {
+	level: number;
+	onClick?: (...args: any) => void;
+};
+
+export function SaturationThresholdButton({ level, onClick }: ThresholdButtonProps) {
+	return (
+		<S.IconWrapper onClick={() => onClick?.()}>
+			<S.FilterIcon saturation={level} />
+		</S.IconWrapper>
+	);
+}
+
+type DirectionButtonProps = {
+	direction: Direction | "any";
+	onClick?: (...args: any) => void;
+};
+
+export function DirectionButton({ direction, onClick }: DirectionButtonProps) {
+	return (
+		<S.DirectionButton
+			key={`saturation-filter-${direction}`}
+			value={direction}
+			onClick={() => onClick?.(direction)}
+		>
+			<span>{direction}</span>
+			{direction !== "any" && (
+				<span>
+					{direction === "forwards" ? <BiArrowToRight /> : <BiArrowToLeft />}
+				</span>
+			)}
+		</S.DirectionButton>
 	);
 }
