+++ frontend/src/components/SaturationFilter/hooks/useSaturationFilter.tsx
@@ -1,19 +1,14 @@
 import { colorMap } from "helpers/list.api";
 import { useState } from "react";
-import { BiArrowToLeft, BiArrowToRight } from "react-icons/bi";
 import { useRecoilState, useResetRecoilState } from "recoil";
 import { termFilterState } from "../../../state/filter";
-import OperatorButton, { Operator } from "../FilterButtons";
+import OperatorButton, {
+	DirectionButton,
+	Operator,
+	SaturationThresholdButton,
+} from "../FilterButtons";
 import { operators } from "../helpers/operators";
-import * as S from "../SaturationFilter.style";
-import { FilterUpdate } from "../types/filter-types";
-
-export enum FilterStep {
-	INITIAL = 0,
-	DIRECTION = 1,
-	OPERATOR = 2,
-	LEVEL = 3,
-}
+import { FilterStep, FilterUpdate } from "../types/filter-types";
 
 export function useSaturationFilter() {
 	const [termFilter, setTermFilter] = useRecoilState(termFilterState);
@@ -37,30 +32,26 @@ export function useSaturationFilter() {
 		cycleFilterStep();
 	}
 
-	const thresholdButtons = Object.keys(colorMap).map((level) => {
-		return (
-			<SaturationThresholdButton
-				level={+level}
-				key={`saturation-wrapper-${level}`}
-				onClick={() => {
-					handleFilterButtonClick({ field: "value", value: +level });
-				}}
-			/>
-		);
-	});
+	const thresholdButtons = Object.keys(colorMap).map((level) => (
+		<SaturationThresholdButton
+			level={+level}
+			key={`saturation-wrapper-${level}`}
+			onClick={() => {
+				handleFilterButtonClick({ field: "value", value: +level });
+			}}
+		/>
+	));
 
 	const directionButtons = ["forwards", "backwards", "any"].map(
-		(direction: Direction | "any") => {
-			return (
-				<DirectionButton
-					key={direction}
-					direction={direction}
-					onClick={() =>
-						handleFilterButtonClick({ field: "direction", value: direction })
-					}
-				/>
-			);
-		}
+		(direction: Direction | "any") => (
+			<DirectionButton
+				key={direction}
+				direction={direction}
+				onClick={() =>
+					handleFilterButtonClick({ field: "direction", value: direction })
+				}
+			/>
+		)
 	);
 
 	const operatorButtons = Object.values(operators).map((operator: Operator) => (
@@ -80,39 +71,4 @@ export function useSaturationFilter() {
 		thresholdButtons,
 		directionButtons,
 	} as const;
-}
-
-type ThresholdButtonProps = {
-	level: number;
-	onClick?: (...args: any) => void;
-};
-
-function SaturationThresholdButton({ level, onClick }: ThresholdButtonProps) {
-	return (
-		<S.IconWrapper onClick={() => onClick?.()}>
-			<S.FilterIcon saturation={level} />
-		</S.IconWrapper>
-	);
-}
-
-type DirectionButtonProps = {
-	direction: Direction | "any";
-	onClick?: (...args: any) => void;
-};
-
-export function DirectionButton({ direction, onClick }: DirectionButtonProps) {
-	return (
-		<S.DirectionButton
-			key={`saturation-filter-${direction}`}
-			value={direction}
-			onClick={() => onClick?.(direction)}
-		>
-			<span>{direction}</span>
-			{direction !== "any" && (
-				<span>
-					{direction === "forwards" ? <BiArrowToRight /> : <BiArrowToLeft />}
-				</span>
-			)}
-		</S.DirectionButton>
-	);
 }
